+++ test/jobs/workflow_executions/workflow_execution_status_job_test.rb
@@ -1,0 +1,518 @@
+# frozen_string_literal: true
+
+require 'active_job/continuation/test_helper'
+require 'test_helper'
+require 'test_helpers/faraday_test_helpers'
+require 'webmock/minitest'
+
+module WorkflowExecutions
+  class WorkflowExecutionStatusJobTest < ActiveJob::TestCase
+    include FaradayTestHelpers
+    include ActiveJob::Continuation::TestHelper
+
+    def setup # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
+      @original_clone_repo_method = Irida::PipelineRepository.method(:clone_repo)
+      @workflow_execution = workflow_executions(:irida_next_example_submitted)
+      @workflow_execution.email_notification = true
+      @workflow_execution.save!
+      @workflow_execution.create_logidze_snapshot!
+      @stubs = faraday_test_adapter_stubs
+
+      @test_schema_body = Rails.root.join('test/fixtures/files/nextflow/nextflow_schema.json').read
+
+      # Use a closure to capture @test_schema_body for the clone_repo implementation
+      schema_body = @test_schema_body
+      clone_repo_impl = lambda do |_uri, _sha, clone_dir|
+        FileUtils.mkdir_p(clone_dir)
+        File.write(File.join(clone_dir, 'nextflow_schema.json'), schema_body)
+
+        # Create assets/schema_input.json
+        FileUtils.mkdir_p(File.join(clone_dir, 'assets'))
+        File.write(File.join(clone_dir, 'assets', 'schema_input.json'), schema_body)
+        nil
+      end
+
+      Irida::PipelineRepository.singleton_class.send(:define_method, :clone_repo, clone_repo_impl)
+    end
+
+    def teardown
+      # reset connections after each test to clear cache
+      Faraday.default_connection = nil
+      Irida::PipelineRepository.singleton_class.send(:define_method, :clone_repo, @original_clone_repo_method)
+    end
+
+    test 'successful job execution' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+      assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+      assert_performed_jobs 1
+      assert @workflow_execution.reload.completing?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'job execution still queued after status check' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'QUEUED'
+            }
+          ]
+        end
+
+        WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert @workflow_execution.reload.submitted?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'job execution from queued to complete' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 4)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'QUEUED'
+            }
+          ]
+        end
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'QUEUED'
+            }
+          ]
+        end
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'RUNNING'
+            }
+          ]
+        end
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_performed_jobs(4, only: WorkflowExecutionStatusJob)
+      assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+      assert @workflow_execution.reload.completing?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'repeated connection errors' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 6)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = "/runs/#{@workflow_execution.run_id}/status"
+        @stubs.get(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.get(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.get(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.get(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.get(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        # a success after 5 failed attempts
+        @stubs.get(endpoint) do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: [WorkflowExecutionStatusJob]) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_performed_jobs(6, only: WorkflowExecutionStatusJob)
+      assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+      assert @workflow_execution.reload.completing?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'repeated api exception errors' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 3)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = "/runs/#{@workflow_execution.run_id}/status"
+        @stubs.get(endpoint) { |_env| raise Faraday::BadRequestError }
+        @stubs.get(endpoint) { |_env| raise Faraday::BadRequestError }
+        @stubs.get(endpoint) { |_env| raise Faraday::BadRequestError }
+        # a success (never reached) after 3 failed attempts
+        @stubs.get(endpoint) do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      # 3 failures
+      assert_performed_jobs(3, only: WorkflowExecutionStatusJob)
+      @workflow_execution.reload
+      assert @workflow_execution.error?
+      assert @workflow_execution.http_error_code == 400
+
+      assert_enqueued_emails 1
+      assert_enqueued_email_with PipelineMailer, :error_user_email, args: [@workflow_execution]
+    end
+
+    test 'api exception error then a success' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 2)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = "/runs/#{@workflow_execution.run_id}/status"
+        @stubs.get(endpoint) { |_env| raise Faraday::BadRequestError }
+        # a success after 1 failed attempt
+        @stubs.get(endpoint) do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: [WorkflowExecutionStatusJob]) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+      assert_performed_jobs(2, only: WorkflowExecutionStatusJob)
+      assert @workflow_execution.reload.completing?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'system error job execution' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'SYSTEM_ERROR'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      assert_performed_jobs 1
+      assert @workflow_execution.reload.error?
+
+      assert_enqueued_emails 1
+      assert_enqueued_email_with PipelineMailer, :error_user_email, args: [@workflow_execution]
+    end
+
+    test 'execution where namespace is removed before status is run' do
+      workflow_execution = workflow_executions(:workflow_execution_missing_namespace)
+      workflow_execution.create_logidze_snapshot!
+
+      perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+        WorkflowExecutionStatusJob.perform_later(workflow_execution)
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert workflow_execution.reload.error?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'execution where run_id is missing' do
+      workflow_execution = workflow_executions(:workflow_execution_missing_run_id)
+      workflow_execution.create_logidze_snapshot!
+
+      perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+        WorkflowExecutionStatusJob.perform_later(workflow_execution)
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert workflow_execution.reload.error?
+    end
+
+    test 'already canceling workflow should return early' do
+      workflow_execution = workflow_executions(:irida_next_example_canceling)
+
+      WorkflowExecutionStatusJob.perform_later(workflow_execution)
+      assert_enqueued_jobs(1, except: Turbo::Streams::BroadcastStreamJob)
+      perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_enqueued_jobs(0, except: Turbo::Streams::BroadcastStreamJob)
+    end
+
+    test 'already canceled workflow should return early' do
+      workflow_execution = workflow_executions(:irida_next_example_canceled)
+      workflow_execution.create_logidze_snapshot!
+
+      WorkflowExecutionStatusJob.perform_later(workflow_execution)
+      assert_enqueued_jobs(1, except: Turbo::Streams::BroadcastStreamJob)
+      perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_enqueued_jobs(0, except: Turbo::Streams::BroadcastStreamJob)
+    end
+
+    test 'successful job execution service returned canceled' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'CANCELED'
+            }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      assert @workflow_execution.reload.canceled?
+
+      assert_no_enqueued_emails
+    end
+
+    test 'min_run_time with running state should delay status check' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+      @workflow_execution.create_logidze_snapshot!
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'RUNNING'
+            }
+          ]
+        end
+
+        assert_enqueued_with(job: WorkflowExecutionStatusJob,
+                             at: Time.current + @workflow_execution.workflow.minimum_run_time) do
+          WorkflowExecutionStatusJob.set(
+            wait_until: @workflow_execution.workflow.minimum_run_time.seconds.from_now
+          ).perform_later(@workflow_execution)
+        end
+
+        assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+        assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      end
+    end
+
+    test 'min_run_time with non-running state should proceed normally' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        assert_enqueued_with(job: WorkflowExecutionStatusJob,
+                             at: Time.current + @workflow_execution.workflow.status_check_interval) do
+          WorkflowExecutionStatusJob.set(
+            wait_until: @workflow_execution.workflow.status_check_interval.seconds.from_now
+          ).perform_later(@workflow_execution)
+        end
+
+        assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+          WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+      assert_performed_jobs(1, only: WorkflowExecutionStatusJob)
+      assert @workflow_execution.reload.completing?
+    end
+
+    test 'workflow execution exceeds maximum runtime should queue cancellation job' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+      # Set up pipeline schema to get max_runtime
+      @pipeline_schema_file_dir = "#{ActiveStorage::Blob.service.root}/pipelines"
+      Irida::Pipelines.new(pipeline_config_file: 'test/config/pipelines/pipelines.json',
+                           pipeline_schema_file_dir: @pipeline_schema_file_dir)
+
+      workflow_execution = workflow_executions(:irida_next_example_running)
+      workflow_execution.create_logidze_snapshot!
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'RUNNING'
+            }
+          ]
+        end
+
+        # Mock state_time_calculation to return a large run time (400 seconds)
+        workflow_execution.workflow.stub :state_time_calculation, 400 do
+          perform_enqueued_jobs(only: WorkflowExecutionStatusJob) do
+            WorkflowExecutionStatusJob.perform_later(workflow_execution)
+          end
+        end
+      end
+
+      assert workflow_execution.reload.canceling?
+      assert_enqueued_jobs(1, only: WorkflowExecutionCancelationJob)
+    end
+
+    test 'workflow execution which hasn\'t exceeded maximum runtime should proceed normally' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+      # Set up pipeline schema to get max_runtime
+      @pipeline_schema_file_dir = "#{ActiveStorage::Blob.service.root}/pipelines"
+      Irida::Pipelines.new(pipeline_config_file: 'test/config/pipelines/pipelines.json',
+                           pipeline_schema_file_dir: @pipeline_schema_file_dir)
+
+      workflow_execution = workflow_executions(:irida_next_example_running)
+      workflow_execution.create_logidze_snapshot!
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'RUNNING'
+            }
+          ]
+        end
+
+        WorkflowExecutionStatusJob.perform_later(workflow_execution)
+        assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+      end
+
+      assert workflow_execution.reload.running?
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+    end
+
+    # Example test case using RSpec/TestUnit
+    test 'status job resumes from the correct step' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.get("/runs/#{@workflow_execution.run_id}/status") do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            {
+              run_id: @workflow_execution.run_id,
+              state: 'COMPLETE'
+            }
+          ]
+        end
+
+        WorkflowExecutionStatusJob.perform_later(@workflow_execution)
+
+        interrupt_job_after_step(WorkflowExecutionStatusJob, :query_and_update_state) do
+          perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+        end
+
+        assert_equal :completing, @workflow_execution.reload.state.to_sym
+        assert_enqueued_jobs(0, only: WorkflowExecutionCompletionJob)
+        assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+
+        perform_enqueued_jobs(only: WorkflowExecutionStatusJob)
+
+        assert_equal :completing, @workflow_execution.reload.state.to_sym
+        assert_enqueued_jobs(1, only: WorkflowExecutionCompletionJob)
+        assert_enqueued_jobs(0, only: WorkflowExecutionStatusJob)
+      end
+    end
+  end
+end
