+++ test/jobs/workflow_executions/workflow_execution_submission_job_test.rb
@@ -1,0 +1,195 @@
+# frozen_string_literal: true
+
+require 'active_job/continuation/test_helper'
+require 'test_helper'
+require 'test_helpers/faraday_test_helpers'
+
+module WorkflowExecutions
+  class WorkflowExecutionSubmissionJobTest < ActiveJob::TestCase
+    include FaradayTestHelpers
+    include ActiveJob::Continuation::TestHelper
+
+    def setup
+      @workflow_execution = workflow_executions(:irida_next_example_prepared)
+      @stubs = faraday_test_adapter_stubs
+    end
+
+    def teardown
+      # reset connections after each test to clear cache
+      Faraday.default_connection = nil
+    end
+
+    test 'successful job execution' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+      test_run_id = '123'
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.post('/runs') do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            { run_id: test_run_id }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+          WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        end
+        @workflow_execution.reload
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(1, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.submitted?
+      assert_equal test_run_id, @workflow_execution.run_id
+    end
+
+    test 'successful job execution with interrupt' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+      test_run_id = '123'
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.post('/runs') do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            { run_id: test_run_id }
+          ]
+        end
+
+        WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        interrupt_job_after_step(WorkflowExecutionSubmissionJob, :query_and_update_state) do
+          perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob)
+        end
+        @workflow_execution.reload
+      end
+
+      assert_enqueued_jobs(0, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(1, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.submitted?
+      assert_equal test_run_id, @workflow_execution.run_id
+
+      perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob)
+      @workflow_execution.reload
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(2, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.submitted?
+      assert_equal test_run_id, @workflow_execution.run_id
+    end
+
+    test 'repeated connection errors' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 6)
+      test_run_id = '123'
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = '/runs'
+        @stubs.post(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.post(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.post(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.post(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        @stubs.post(endpoint) { |_env| raise Faraday::ConnectionFailed }
+        # a success after 5 failed attempts
+        @stubs.post(endpoint) do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            { run_id: test_run_id }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+          WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(6, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.reload.submitted?
+    end
+
+    test 'repeated api exception errors' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 3)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = '/runs'
+        @stubs.post(endpoint) { |_env| raise Faraday::BadRequestError }
+        @stubs.post(endpoint) { |_env| raise Faraday::BadRequestError }
+        @stubs.post(endpoint) { |_env| raise Faraday::BadRequestError }
+
+        perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+          WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionCleanupJob)
+      assert_performed_jobs(3, only: WorkflowExecutionSubmissionJob)
+      @workflow_execution.reload
+      assert @workflow_execution.error?
+      assert @workflow_execution.http_error_code == 400
+    end
+
+    test 'api exception error then a success' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 2)
+      test_run_id = '123'
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        endpoint = '/runs'
+        @stubs.post(endpoint) { |_env| raise Faraday::BadRequestError }
+        # a success after 1 failed attempt
+        @stubs.post(endpoint) do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            { run_id: test_run_id }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+          WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(2, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.reload.submitted?
+    end
+
+    test 'not completing job execution' do
+      workflow_execution = workflow_executions(:irida_next_example_submitted)
+
+      perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+        WorkflowExecutionSubmissionJob.perform_later(workflow_execution)
+      end
+
+      assert_enqueued_jobs(0, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(1, only: WorkflowExecutionSubmissionJob)
+      assert_not workflow_execution.reload.submitted?
+      assert workflow_execution.error?
+    end
+
+    # This test is a unique case where the error is expected to be caught by the status job at a later point
+    test 'job execution succeeded despite WES returning nil run_id' do
+      mock_client = connection_builder(stubs: @stubs, connection_count: 1)
+
+      Integrations::Ga4ghWesApi::V1::ApiConnection.stub :new, mock_client do
+        @stubs.post('/runs') do |_env|
+          [
+            200,
+            { 'Content-Type': 'application/json' },
+            { run_id: nil }
+          ]
+        end
+
+        perform_enqueued_jobs(only: WorkflowExecutionSubmissionJob) do
+          WorkflowExecutionSubmissionJob.perform_later(@workflow_execution)
+        end
+      end
+
+      assert_enqueued_jobs(1, only: WorkflowExecutionStatusJob)
+      assert_performed_jobs(1, only: WorkflowExecutionSubmissionJob)
+      assert @workflow_execution.reload.submitted?
+      assert_not @workflow_execution.error?
+    end
+  end
+end
